[debug]update product

+++ app/controllers/updateProduct.js
@@ -18,7 +18,7 @@ module.exports = async (ctx) => {
     } = _params;
 
     // 更新图片列表
-    if(!!pro_image_list){
+    if(!!pro_image_list && !!pro_id){
         await ProductImageInfo.destroy({
             where:{
                 pro_id: pro_id,
@@ -39,15 +39,15 @@ module.exports = async (ctx) => {
 
 
     // 更新详情列表
-    if(!!pro_detail_list) {
-        ProductDetailInfo.destroy({
+    if(!!pro_detail_list && !!pro_id) {
+        await ProductDetailInfo.destroy({
             where:{
                 pro_id: pro_id,
             }
         })
 
         const detail_list = JSON.parse(pro_detail_list).list;
-        !!detail_list && detail_list.length>0 && detail_list.map((detail,idx) => {
+        !!detail_list && detail_list.length>0 && await detail_list.map((detail,idx) => {
             ProductDetailInfo.create({
                 ...detail,
                 pro_id: pro_id,
@@ -57,15 +57,15 @@ module.exports = async (ctx) => {
     }
 
     // 更新价格列表
-    if(!!price_list) {
-        ProductPriceInfo.destroy({
+    if(!!price_list && !!pro_id) {
+        await ProductPriceInfo.destroy({
             where:{
                 pro_id: pro_id,
             }
         })
 
         const _price_list = JSON.parse(price_list).list;
-        !!_price_list && _price_list.length>0 && _price_list.map((price,idx) => {
+        !!_price_list && _price_list.length>0 && await _price_list.map((price,idx) => {
             ProductPriceInfo.create({
                 ...price,
                 pro_id: pro_id,
@@ -81,7 +81,6 @@ module.exports = async (ctx) => {
             pro_id: pro_id,
         }
     })
-    console.log('_params',_params);
 
     ctx.body = {'message':'update product success'};
 }
